refactor(patterns): Split pattern detection into modular components
