+++ components/RecurringTasks/utils/getTaskFrequency.ts
@@ -12,20 +12,23 @@ export function getTaskFrequency(dueString: string | undefined): RecurringFreque
     /every \d+ days?/.test(lower)
   ) return 'daily';
 
-  if (lower.includes('every other')) return 'other';
+  // Check for weekly patterns, including biweekly
+  if (
+    lower.includes('every week') || 
+    lower.includes('weekly') ||
+    /every (other )?((monday|tuesday|wednesday|thursday|friday|saturday|sunday)|(mon|tue|wed|thu|fri|sat|sun))/i.test(lower)
+  ) return 'weekly';
   
+  // Check for monthly patterns
   if (
     lower.includes('every') && lower.includes('month') ||
     lower.includes('monthly') ||
     /every (\d{1,2})(st|nd|rd|th)?( day)?( of)?( the)?( month)?$/i.test(lower) || 
     /every (first|last|1st) day/i.test(lower)
   ) return 'monthly';
-  if (
-    lower.includes('every week') || 
-    lower.includes('weekly') ||
-    /every (monday|tuesday|wednesday|thursday|friday|saturday|sunday)/i.test(lower) ||
-    /every (mon|tue|wed|thu|fri|sat|sun)/i.test(lower)
-  ) return 'weekly';
+
+  // Catch any remaining "every other" patterns
+  if (lower.includes('every other')) return 'other';
   
   return 'other';
 }
